Correct resource closing logic
Make the objects that open resources be responsible for closing said opened resources. It is not the responsibility of objects themselves to do their own clean up. Additionally, if I hand a child resource to some method I do not it expect to be returned in a closed, unusable state barring some error.